Added teacher name retrieval function
This allows other functions to get the name of a teacher when given a class number. Either int or string classNums work, since when a class is loaded it's converted to a int, but when a class is retrieved directly from a file, it's a string, hence the extra compatibility code.